refactor: change the method of panel animation
before panels used to be reveal animated by just changing the position.
hidden would be away from the current screen, or under it or whatever.

now it is at a static position, but is instead shaped with the animated
width. this performs better and looks better, but might be odd when
i want panels to actually look like they're moving (like the brief notif, or
maybe panels like volume)